Fix media port in sipp_uas_pcap_g711a_multi_183.xml
sipp has an internal data structure play_args_t to play pcap files. This
structure is not populated correctly with the local media port if the [media_port]
keyword is not explicitly used within the SDP, and therefore UDP packets start being
sent from source port 0, which at least in my tests caused the packets to not be
received at all in the other application, not sure if dropped in the kernel stack
or discarded upon reception. Overall is a better idea to use [media_port] in the sipp
scenarios that use pcap and adjust the port using the -mp option for sipp